ファイル分割とbugfix
